split numerical_basis.cpp (2/2): extract output_info+output_k+output_overlap_Q/Sq/V to numerical_basis_output.cpp
